Small changes to figure sizes for random initial guesses
I trimmed the .pdfs but it will cause small differences in the figures. Overall, I'm not sure trimming is a good idea.

We can make a single figure out of the four figures in Illustrator. That's a possibility. However, the new Xs and red stars actually work quite well. Good job Qiushi.